Move gridsample from grid method to standalone function; combine all sampling functions into single file (sampling.py)
